Count the leaves of the declaration-versus-statement case in the width the expectation is compared at, a widening cast applied after the arithmetic having drawn two lint errors
The case computed each leaf's expected value in int and cast the result, which is the shape one check reads as either ineffective or lossy, the widening having come after the multiplication rather than before it, and which another reads as a type name written twice where the initializer already says it. Both go when the loop counters carry the width themselves: the arithmetic is then done in the width the comparison uses, and the cast that provoked the pair is gone rather than moved.

Nothing about what the case asserts changes; the six values are the same six.

Refs #3577

+++ test/src/unit/test_simulator_subclause_10_09_01.cpp
@@ -700,14 +700,14 @@ TEST(ArrayLiteralSim, VarInitLeafMatchesProceduralAssignLeaf) {
       f);
   ASSERT_NE(design, nullptr);
   LowerAndRun(design, f);
-  for (int i = 1; i <= 2; ++i) {
-    for (int j = 1; j <= 3; ++j) {
+  for (uint64_t i = 1; i <= 2; ++i) {
+    for (uint64_t j = 1; j <= 3; ++j) {
       auto suffix = "[" + std::to_string(i) + "][" + std::to_string(j) + "]";
       auto* declared = f.ctx.FindVariable("d" + suffix);
       auto* assigned = f.ctx.FindVariable("p" + suffix);
       ASSERT_NE(declared, nullptr) << suffix;
       ASSERT_NE(assigned, nullptr) << suffix;
-      uint64_t want = static_cast<uint64_t>(((i - 1) * 3 + j) * 10);
+      uint64_t want = ((i - 1) * 3 + j) * 10;
       EXPECT_EQ(assigned->value.ToUint64(), want) << suffix;
       EXPECT_EQ(declared->value.ToUint64(), want) << suffix;
     }
